backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "typing>=3.10.0.0",
    "uvicorn>=0.40.0",
]
